Windows: fix compiler warnings

+++ libcuckoo/cuckoohash_map.hh
@@ -330,7 +330,7 @@ private:
     static size_t reserve_calc(const size_t n) {
         const size_t buckets = (n + slot_per_bucket - 1) / slot_per_bucket;
         size_t blog2;
-        for (blog2 = 1; (1UL << blog2) < buckets; ++blog2);
+        for (blog2 = 1; (1ULL << blog2) < buckets; ++blog2);
         assert(n <= hashsize(blog2) * slot_per_bucket);
         return blog2;
     }
@@ -684,7 +684,7 @@ private:
         const uint16_t hash_16bit = (
             static_cast<uint16_t>(hash_32bit) ^
             static_cast<uint16_t>(hash_32bit >> 16));
-        const uint16_t hash_8bit = (
+        const uint8_t hash_8bit = (
             static_cast<uint8_t>(hash_16bit) ^
             static_cast<uint8_t>(hash_16bit >> 8));
         return hash_8bit;
